+++ example/example_streamed_request.dart
@@ -54,13 +54,20 @@ void main() async {
   final token = cts.token;
   final client = CancelableClient(token);
   try {
-    final request = MultipartRequest("POST", url);
+    final request = StreamedRequest("POST", url);
     final file = File(filepath);
     // Make it possible to cancel sending data.
     final stream = file.openRead().asCancelable(token, throwIfCanceled: true);
-    request.files.add(MultipartFile('file', stream, file.lengthSync()));
+    final sink = request.sink;
     request.headers['Content-Type'] = 'text/plain';
-    _client('Sending multipart request with timeout $timeout ms');
+    request.contentLength = file.lengthSync();
+    stream.listen(
+      sink.add,
+      onDone: sink.close,
+      onError: sink.addError,
+      cancelOnError: true,
+    );
+    _client('Sending streaming request with timeout $timeout ms');
     await client.send(request);
     cts.cancelAfter(null);
     _client('Received response');